Comment out header addition
I'm temporarily removing this because the outputted dats are expected to behave like the old ones. This means that each of the systems will have headered AND unheadered variants. WIth the skips, then only the unheadered variant is going to be detected properly. In the mean time, this is being removed so that sets can be properly created for sharing.